refactor(cnc): regrouped cnc ops into a single module, updated dependants
